Режимы flow работают, но: 1. Нет такого что пока 1 шаровый закрывается другой открывается 2. Чистка это непонятная фигня, работает как написал(в том как написал логику не уверен, логика: все магитрали( открываем шар клапан актуатор) 10 сек, потом продуваемм 22 сек)

+++ types_for_magistral.hpp
@@ -1,7 +1,7 @@
 #ifndef TYPES_FOR_MAGISTRAL_HPP
 #define TYPES_FOR_MAGISTRAL_HPP
 
-typedef int Time_in_this;
+typedef uint16_t Time_in_this;
 
 enum state_Magistral {
     start_state = 0,
